Split ImageWMS and TileWMS layers heralds into different files
