+++ src/herald/layersherald.js
@@ -3,8 +3,9 @@ goog.provide('olgm.herald.Layers');
 goog.require('goog.asserts');
 goog.require('olgm');
 goog.require('olgm.gm');
-goog.require('olgm.gm.ImageOverlay');
 goog.require('olgm.herald.Herald');
+goog.require('olgm.herald.ImageWMSSource');
+goog.require('olgm.herald.TileWMSSource');
 goog.require('olgm.herald.VectorSource');
 goog.require('olgm.herald.View');
 goog.require('olgm.layer.Google');
@@ -74,28 +75,16 @@ olgm.herald.Layers = function(ol3map, gmap, watchVector) {
   this.vectorLayers_ = [];
 
   /**
-   * @type {Array.<olgm.herald.Layers.ImageWMSLayerCache>}
+   * @type {olgm.herald.ImageWMSSource}
    * @private
    */
-  this.imageWMSCache_ = [];
+  this.imageWMSSourceHerald_ = new olgm.herald.ImageWMSSource(ol3map, gmap);
 
   /**
-   * @type {Array.<ol.layer.Image>}
+   * @type {olgm.herald.TileWMSSource}
    * @private
    */
-  this.imageWMSLayers_ = [];
-
-  /**
-   * @type {Array.<olgm.herald.Layers.TileWMSLayerCache>}
-   * @private
-   */
-  this.tileWMSCache_ = [];
-
-  /**
-   * @type {Array.<ol.layer.Tile>}
-   * @private
-   */
-  this.tileWMSLayers_ = [];
+  this.tileWMSSourceHerald_ = new olgm.herald.TileWMSSource(ol3map, gmap);
 
   /**
    * @type {olgm.herald.View}
@@ -210,6 +199,18 @@ olgm.herald.Layers.prototype.getGoogleMapsActive = function() {
 
 
 /**
+ * Set the googleMapsIsActive value and spread the change to the heralds
+ * @param {boolean} active
+ * @private
+ */
+olgm.herald.Layers.prototype.setGoogleMapsActive_ = function(active) {
+  this.googleMapsIsActive_ = active;
+  this.imageWMSSourceHerald_.setGoogleMapsActive(active);
+  this.tileWMSSourceHerald_.setGoogleMapsActive(active);
+};
+
+
+/**
  * Callback method fired when a new layer is added to the map.
  * @param {ol.CollectionEvent} event Collection event.
  * @private
@@ -246,12 +247,12 @@ olgm.herald.Layers.prototype.watchLayer_ = function(layer) {
   } else if (layer instanceof ol.layer.Tile) {
     var source = layer.getSource();
     if (source instanceof ol.source.TileWMS) {
-      this.watchTileWMSLayer_(layer);
+      this.tileWMSSourceHerald_.watchLayer(layer);
     }
   } else if (layer instanceof ol.layer.Image) {
     var source = layer.getSource();
     if (source instanceof ol.source.ImageWMS) {
-      this.watchImageWMSLayer_(layer);
+      this.imageWMSSourceHerald_.watchLayer(layer);
     }
   }
 };
@@ -271,152 +272,6 @@ olgm.herald.Layers.prototype.watchGoogleLayer_ = function(layer) {
     ]
   }));
   this.toggleGoogleMaps_();
-};
-
-
-/**
- * Watch an image WMS layer, and create a layer on the google maps layer
- * @param {ol.layer.Image} layer
- * @private
- */
-olgm.herald.Layers.prototype.watchImageWMSLayer_ = function(layer) {
-  // Source required
-  var source = layer.getSource();
-  if (!source) {
-    return;
-  }
-
-  this.imageWMSLayers_.push(layer);
-
-  // opacity
-  var opacity = layer.getOpacity();
-
-  var cacheItem = /** {@type olgm.herald.Layers.ImageWMSLayerCache} */ ({
-    imageOverlay: null,
-    lastUrl: null,
-    layer: layer,
-    listenerKeys: [],
-    opacity: opacity
-  });
-
-  // Hide the google layer when the ol3 layer is invisible
-  cacheItem.listenerKeys.push(layer.on('change:visible',
-      this.handleImageWMSLayerVisibleChange_.bind(this, cacheItem), this));
-
-  cacheItem.listenerKeys.push(this.ol3map.on('moveend',
-      this.handleImageWMSMoveEnd_.bind(this, cacheItem), this));
-
-  cacheItem.listenerKeys.push(this.ol3map.getView().on('change:resolution',
-      this.handleImageWMSMoveEnd_.bind(this, cacheItem), this));
-
-  // Activate the cache item
-  this.activateImageWMSLayerCacheItem_(cacheItem);
-  this.imageWMSCache_.push(cacheItem);
-};
-
-
-/**
- * Generate a wms request url for a single image
- * @param {ol.layer.Image} layer
- * @return {string}
- * @private
- */
-olgm.herald.Layers.prototype.generateImageWMSFunction_ = function(layer) {
-  var source = layer.getSource();
-  goog.asserts.assertInstanceof(source, ol.source.ImageWMS);
-
-  var params = source.getParams();
-  var ol3map = this.ol3map;
-
-  //base WMS URL
-  var url = source.getUrl();
-  var size = ol3map.getSize();
-
-  goog.asserts.assert(size !== undefined);
-
-  var view = ol3map.getView();
-  var bbox = view.calculateExtent(size);
-
-  // Get params
-  var version = params['VERSION'] ? params['VERSION'] : '1.3.0';
-  var layers = params['LAYERS'] ? params['LAYERS'] : '';
-  var styles = params['STYLES'] ? params['STYLES'] : '';
-  var format = params['FORMAT'] ? params['FORMAT'] : 'image/png';
-  var transparent = params['TRANSPARENT'] ? params['TRANSPARENT'] : 'TRUE';
-  var tiled = params['TILED'] ? params['TILED'] : 'FALSE';
-
-  url += '?SERVICE=WMS';
-  url += '&VERSION=' + version;
-  url += '&REQUEST=GetMap';
-  url += '&LAYERS=' + layers;
-  url += '&STYLES=' + styles;
-  url += '&FORMAT=' + format;
-  url += '&TRANSPARENT=' + transparent;
-  url += '&SRS=EPSG:3857';
-  url += '&BBOX=' + bbox;
-  url += '&WIDTH=' + size[0];
-  url += '&HEIGHT=' + size[1];
-  url += '&TILED=' + tiled;
-
-  return url;
-};
-
-
-/**
- * Watch a tiled WMS layer, and create a layer on the google maps layer
- * @param {ol.layer.Tile} layer
- * @private
- */
-olgm.herald.Layers.prototype.watchTileWMSLayer_ = function(layer) {
-  // Source required
-  var source = layer.getSource();
-  if (!source) {
-    return;
-  }
-  goog.asserts.assertInstanceof(source, ol.source.TileWMS);
-
-  this.tileWMSLayers_.push(layer);
-
-  // opacity
-  var opacity = layer.getOpacity();
-
-  var cacheItem = /** {@type olgm.herald.Layers.TileWMSLayerCache} */ ({
-    googleWMSLayer: null,
-    layer: layer,
-    listenerKeys: [],
-    opacity: opacity
-  });
-
-  var getTileUrlFunction = source.getTileUrlFunction();
-  var proj = ol.proj.get('EPSG:3857');
-
-  var googleGetTileUrlFunction = function(coords, zoom) {
-    var ol3Coords = [zoom, coords.x, (-coords.y) - 1];
-    return getTileUrlFunction(ol3Coords, 1, proj);
-  };
-
-  var tileSize = new google.maps.Size(256, 256);
-
-  var options = {
-    'getTileUrl': googleGetTileUrlFunction,
-    'tileSize': tileSize,
-    'isPng': true
-  };
-
-  // Create the WMS layer on the google layer
-  var googleWMSLayer = new google.maps.ImageMapType(options);
-  if (layer.getVisible()) {
-    this.gmap.overlayMapTypes.push(googleWMSLayer);
-  }
-  cacheItem.googleWMSLayer = googleWMSLayer;
-
-  // Hide the google layer when the ol3 layer is invisible
-  cacheItem.listenerKeys.push(layer.on('change:visible',
-      this.handleTileWMSLayerVisibleChange_.bind(this, cacheItem), this));
-
-  // Activate the cache item
-  this.activateTileWMSLayerCacheItem_(cacheItem);
-  this.tileWMSCache_.push(cacheItem);
 };
 
 
@@ -485,12 +340,12 @@ olgm.herald.Layers.prototype.unwatchLayer_ = function(layer) {
   } else if (layer instanceof ol.layer.Tile) {
     var source = layer.getSource();
     if (source instanceof ol.source.TileWMS) {
-      this.unwatchTileWMSLayer_(layer);
+      this.tileWMSSourceHerald_.unwatchLayer(layer);
     }
   } else if (layer instanceof ol.layer.Image) {
     var source = layer.getSource();
     if (source instanceof ol.source.ImageWMS) {
-      this.unwatchImageWMSLayer_(layer);
+      this.imageWMSSourceHerald_.unwatchLayer(layer);
     }
   }
 };
@@ -512,48 +367,6 @@ olgm.herald.Layers.prototype.unwatchGoogleLayer_ = function(layer) {
     this.googleCache_.splice(index, 1);
 
     this.toggleGoogleMaps_();
-  }
-};
-
-
-/**
- * Unwatch the WMS Image layer
- * @param {ol.layer.Image} layer
- * @private
- */
-olgm.herald.Layers.prototype.unwatchImageWMSLayer_ = function(layer) {
-  var index = this.imageWMSLayers_.indexOf(layer);
-  if (index !== -1) {
-    this.imageWMSLayers_.splice(index, 1);
-
-    var cacheItem = this.imageWMSCache_[index];
-    olgm.unlistenAllByKey(cacheItem.listenerKeys);
-
-    // opacity
-    layer.setOpacity(cacheItem.opacity);
-
-    this.imageWMSCache_.splice(index, 1);
-  }
-};
-
-
-/**
- * Unwatch the WMS Tile layer
- * @param {ol.layer.Tile} layer
- * @private
- */
-olgm.herald.Layers.prototype.unwatchTileWMSLayer_ = function(layer) {
-  var index = this.tileWMSLayers_.indexOf(layer);
-  if (index !== -1) {
-    this.tileWMSLayers_.splice(index, 1);
-
-    var cacheItem = this.tileWMSCache_[index];
-    olgm.unlistenAllByKey(cacheItem.listenerKeys);
-
-    // opacity
-    layer.setOpacity(cacheItem.opacity);
-
-    this.tileWMSCache_.splice(index, 1);
   }
 };
 
@@ -614,12 +427,12 @@ olgm.herald.Layers.prototype.activateGoogleMaps_ = function() {
   this.viewHerald_.setCenter();
   this.viewHerald_.setZoom();
 
-  this.googleMapsIsActive_ = true;
+  this.setGoogleMapsActive_(true);
 
   // activate all cache items
   this.vectorCache_.forEach(this.activateVectorLayerCacheItem_, this);
-  this.tileWMSCache_.forEach(this.activateTileWMSLayerCacheItem_, this);
-  this.imageWMSCache_.forEach(this.activateImageWMSLayerCacheItem_, this);
+  this.tileWMSSourceHerald_.activate();
+  this.imageWMSSourceHerald_.activate();
 };
 
 
@@ -645,10 +458,10 @@ olgm.herald.Layers.prototype.deactivateGoogleMaps_ = function() {
 
   // deactivate all cache items
   this.vectorCache_.forEach(this.deactivateVectorLayerCacheItem_, this);
-  this.tileWMSCache_.forEach(this.deactivateTileWMSLayerCacheItem_, this);
-  this.imageWMSCache_.forEach(this.deactivateImageWMSLayerCacheItem_, this);
+  this.tileWMSSourceHerald_.deactivate();
+  this.imageWMSSourceHerald_.deactivate();
 
-  this.googleMapsIsActive_ = false;
+  this.setGoogleMapsActive_(false);
 };
 
 
@@ -699,64 +512,6 @@ olgm.herald.Layers.prototype.toggleGoogleMaps_ = function() {
 
 
 /**
- * Activates an image WMS layer cache item.
- * @param {olgm.herald.Layers.ImageWMSLayerCache} cacheItem
- * @private
- */
-olgm.herald.Layers.prototype.activateImageWMSLayerCacheItem_ = function(
-    cacheItem) {
-  var layer = cacheItem.layer;
-  var visible = layer.getVisible();
-  if (visible && this.googleMapsIsActive_) {
-    cacheItem.lastUrl = null;
-    cacheItem.layer.setOpacity(0);
-    this.updateImageOverlay_(cacheItem);
-  }
-};
-
-
-/**
- * Deactivates an Image WMS layer cache item.
- * @param {olgm.herald.Layers.ImageWMSLayerCache} cacheItem
- * @private
- */
-olgm.herald.Layers.prototype.deactivateImageWMSLayerCacheItem_ = function(
-    cacheItem) {
-  if (cacheItem.imageOverlay) {
-    cacheItem.imageOverlay.setMap(null);
-    cacheItem.imageOverlay = null;
-  }
-  cacheItem.layer.setOpacity(cacheItem.opacity);
-};
-
-
-/**
- * Activates a tiled WMS layer cache item.
- * @param {olgm.herald.Layers.TileWMSLayerCache} cacheItem
- * @private
- */
-olgm.herald.Layers.prototype.activateTileWMSLayerCacheItem_ = function(
-    cacheItem) {
-  var layer = cacheItem.layer;
-  var visible = layer.getVisible();
-  if (visible && this.googleMapsIsActive_) {
-    cacheItem.layer.setOpacity(0);
-  }
-};
-
-
-/**
- * Deactivates a tiled WMS layer cache item.
- * @param {olgm.herald.Layers.TileWMSLayerCache} cacheItem
- * @private
- */
-olgm.herald.Layers.prototype.deactivateTileWMSLayerCacheItem_ = function(
-    cacheItem) {
-  cacheItem.layer.setOpacity(cacheItem.opacity);
-};
-
-
-/**
  * Activates a vector layer cache item, i.e. activate its herald and
  * render the layer invisible. Will only do so if the layer is visible.
  * @param {olgm.herald.Layers.VectorLayerCache} cacheItem
@@ -783,127 +538,6 @@ olgm.herald.Layers.prototype.deactivateVectorLayerCacheItem_ = function(
     cacheItem) {
   cacheItem.herald.deactivate();
   cacheItem.layer.setOpacity(cacheItem.opacity);
-};
-
-
-/**
- * Deal with the google WMS layer when we enable or disable the OL3 WMS layer
- * @param {olgm.herald.Layers.TileWMSLayerCache} cacheItem
- * @private
- */
-olgm.herald.Layers.prototype.handleTileWMSLayerVisibleChange_ = function(
-    cacheItem) {
-  var layer = cacheItem.layer;
-  var visible = layer.getVisible();
-
-  var googleWMSLayer = cacheItem.googleWMSLayer;
-  var googleMapsLayers = this.gmap.overlayMapTypes;
-
-  // Get the position of the google layer so we can remove it
-  var layerIndex = googleMapsLayers.getArray().indexOf(googleWMSLayer);
-
-  if (visible) {
-    // Add the google WMS layer only if it's not there already
-    if (layerIndex == -1) {
-      googleMapsLayers.push(googleWMSLayer);
-    }
-    this.activateTileWMSLayerCacheItem_(cacheItem);
-  } else {
-    // Remove the google WMS layer from the map if it hasn't been done already
-    if (layerIndex != -1) {
-      googleMapsLayers.removeAt(layerIndex);
-    }
-    this.deactivateTileWMSLayerCacheItem_(cacheItem);
-  }
-};
-
-
-/**
- * Deal with the google WMS layer when we enable or disable the OL3 WMS layer
- * @param {olgm.herald.Layers.ImageWMSLayerCache} cacheItem
- * @private
- */
-olgm.herald.Layers.prototype.handleImageWMSLayerVisibleChange_ = function(
-    cacheItem) {
-  var layer = cacheItem.layer;
-  var visible = layer.getVisible();
-
-  if (visible) {
-    this.activateImageWMSLayerCacheItem_(cacheItem);
-  } else {
-    this.deactivateImageWMSLayerCacheItem_(cacheItem);
-  }
-};
-
-
-/**
- * Refresh the custom image overlay on google maps
- * @param {olgm.herald.Layers.ImageWMSLayerCache} cacheItem
- * @private
- */
-olgm.herald.Layers.prototype.updateImageOverlay_ = function(cacheItem) {
-  var layer = cacheItem.layer;
-  var url = this.generateImageWMSFunction_(layer);
-
-  /* We listen to both change:resolution and moveend events. However, changing
-   * resolution eventually sends a moveend event as well. Using only the
-   * moveend event makes zooming in/out look bad. To prevent rendering the
-   * overlay twice when it happens, we save the request url, and if it's the
-   * same as the last time, we don't render it.
-   */
-  if (url == cacheItem.lastUrl) {
-    return;
-  }
-
-  cacheItem.lastUrl = url;
-
-  // Create a new overlay
-  var view = this.ol3map.getView();
-  var size = this.ol3map.getSize();
-
-  goog.asserts.assert(size !== undefined);
-
-  var extent = view.calculateExtent(size);
-
-  // First, get the coordinates of the top left corner
-  var topLeft = ol.extent.getTopLeft(extent);
-
-  // Then, convert it to LatLng coordinates for google
-  var lngLat = ol.proj.transform(topLeft, 'EPSG:3857', 'EPSG:4326');
-  var topLeftLatLng = new google.maps.LatLng(lngLat[1], lngLat[0]);
-
-  var overlay = new olgm.gm.ImageOverlay(
-      url,
-      size,
-      topLeftLatLng);
-
-  // Set the new overlay right away to give it time to render
-  overlay.setMap(this.gmap);
-
-  // Clean previous overlay
-  if (cacheItem.imageOverlay) {
-    // Remove the overlay from the map
-    cacheItem.imageOverlay.setMap(null);
-
-    // Destroy the overlay
-    cacheItem.imageOverlay = null;
-  }
-
-  // Save new overlay
-  cacheItem.imageOverlay = overlay;
-};
-
-
-/**
- * Handle the map being panned when an ImageWMS layer is present
- * @param {olgm.herald.Layers.ImageWMSLayerCache} cacheItem
- * @private
- */
-olgm.herald.Layers.prototype.handleImageWMSMoveEnd_ = function(
-    cacheItem) {
-  if (cacheItem.layer.getVisible()) {
-    this.updateImageOverlay_(cacheItem);
-  }
 };
 
 
@@ -942,26 +576,3 @@ olgm.herald.Layers.GoogleLayerCache;
  * }}
  */
 olgm.herald.Layers.VectorLayerCache;
-
-
-/**
- * @typedef {{
- *   googleWMSLayer: (google.maps.ImageMapType),
- *   layer: (ol.layer.Tile),
- *   listenerKeys: (Array.<ol.events.Key|Array.<ol.events.Key>>),
- *   opacity: (number)
- * }}
- */
-olgm.herald.Layers.TileWMSLayerCache;
-
-
-/**
- * @typedef {{
- *   imageOverlay: (olgm.gm.ImageOverlay),
- *   lastUrl: (string|null),
- *   layer: (ol.layer.Image),
- *   listenerKeys: (Array.<ol.events.Key|Array.<ol.events.Key>>),
- *   opacity: (number)
- * }}
- */
-olgm.herald.Layers.ImageWMSLayerCache;
